rename: Thumbnail 철자 수정
