+++ src/shared/Tab/components/TabThumbnail.tsx
@@ -1,0 +1,11 @@
+import React from 'react';
+
+export interface TabThumbnailProps {
+  children: React.ReactNode;
+}
+
+const TabThumbnail = ({ children }: TabThumbnailProps) => {
+  return children;
+};
+
+export default TabThumbnail;
