--- src/shared/Tab/components/TabThumnail.tsx
@@ -1,11 +1,0 @@
-import React from 'react';
-
-export interface TabThumnailProps {
-  children: React.ReactNode;
-}
-
-const TabThumnail = ({ children }: TabThumnailProps) => {
-  return children;
-};
-
-export default TabThumnail;
