+++ src/shared/Tab/index.ts
@@ -1,3 +1,3 @@
 export { default as Tab } from './components/Tab';
 export { default as TabItem } from './components/TabItem';
-export { default as TabThumnail } from './components/TabThumnail';
+export { default as TabThumbnail } from './components/TabThumbnail';
